refactor(#54): 유저 회원가입 폴더, 파일명 수정
