+++ src/screens/(Common)/EmailCertificate/EmailCertificate.styles.js
