+++ src/screens/(Common)/EmailCertificate/index.js
