+++ src/screens/(Common)/PhoneCertificate/PhoneCertificate.styles.js
