+++ src/screens/(User)/Register/UserRegisterAgree/Agree.styles.js
