+++ src/screens/(User)/Register/UserRegisterInfo/UserInfo.styles.js
