+++ src/screens/(User)/Register/UserRegisterIntro/Intro.styles.js
